Make align a subcommand
